+++ src/main/java/org/pokeherb/orderservice/application/service/dto/response/OrderResponse.java
@@ -1,4 +1,4 @@
-package org.pokeherb.orderservice.presentation.dto;
+package org.pokeherb.orderservice.application.service.dto.response;
 
 import org.pokeherb.orderservice.domain.entity.Order;
 import org.pokeherb.orderservice.domain.entity.OrderStatus;
@@ -6,29 +6,39 @@ import org.pokeherb.orderservice.domain.entity.OrderStatus;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-// 목록용 요약 응답
-public record OrderSummaryResponse(
+// 상세 조회 응답
+public record OrderResponse(
         UUID id,
         UUID productId,
         UUID orderUserId,
         UUID requestVendorId,
+        UUID deliveryDriverId,
         OrderStatus status,
         LocalDateTime dueAt,
+        LocalDateTime cancelledAt,
         LocalDateTime createdAt,
+        LocalDateTime updatedAt,
+        LocalDateTime deletedAt,
         String productName,
-        int quantity
+        int quantity,
+        String requestMemo
 ) {
-    public static OrderSummaryResponse from(Order order) {
-        return new OrderSummaryResponse(
+    public static OrderResponse from(Order order) {
+        return new OrderResponse(
                 order.getId(),
                 order.getProductId(),
                 order.getOrderUserId(),
                 order.getRequestVendorId(),
+                order.getDeliveryDriverId(),
                 order.getOrderStatus(),
                 order.getDueAt(),
+                order.getCancelledAt(),
                 order.getCreatedAt(),
+                order.getUpdatedAt(),
+                order.getDeletedAt(),
                 order.getProductName(),
-                order.getQuantity()
+                order.getQuantity(),
+                order.getRequestMemo()
         );
     }
 }
